+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - **Feature**: Add option to de-duplicate identical files ([#1112](https://github.com/pulsejet/memories/issues/1112)).
+- **Feature**: Show current date at top of timeline ([#1116](https://github.com/pulsejet/memories/issues/1116))
 
 ## [v7.1.0] - 2024-04-01